Simply implementation for notices
Removes the _default attribute of [notice] tags. Now uses the first
emoji it finds in the body of the notice and adds .notice-X to the
enclosing blockquote (where X = the name of the emoji minus the
surrounding colons).

This way, users can use any arbitrary emoji as a notice icon, and target
them with CSS.